Tests: - Move stressActing into a separate test suite - and logging for fastProducer - relax iterations count for fastProducer and stressActing
