+++ krpc/krpc-core/src/commonTest/kotlin/kotlinx/rpc/krpc/KrpcReceiveHandlerTest.kt
@@ -107,101 +107,6 @@ internal class KrpcReceiveHandlerTest : KrpcReceiveHandlerBaseTest() {
         val message2 = (channel.receive() as KrpcCallMessage.CallException).cause.deserialize().message.orEmpty()
         assertTrue(message2.contains("1 messages were unprocessed"), message2)
     }
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    @Test
-    fun stressActing() {
-        val actorJob = Job()
-        val collected = mutableListOf<KrpcMessage>()
-        val bufferSize = stressBufferSize
-
-        runActingTest(
-            callTimeOut = 10.seconds,
-            bufferSize = bufferSize,
-            callHandler = { collected.add(it) },
-            timeout = 360.seconds,
-        ) { acting ->
-            val sendChannel = Channel<KrpcTransportMessage>(Channel.UNLIMITED)
-            val sender = KrpcSendHandler(sendChannel)
-            sender.updateWindowSize(bufferSize)
-
-            val windowJob = launch {
-                while (true) {
-                    val window = when (val message = channel.receive()) {
-                        is KrpcCallMessage.CallException -> fail(
-                            "Unexpected call exception",
-                            message.cause.deserialize()
-                        )
-
-                        is KrpcGenericMessage -> decodeWindow(message)
-                        else -> fail("Unexpected message: $message")
-                    }
-
-                    sender.updateWindowSize((window as WindowResult.Success).update)
-                }
-            }
-
-            val senderJob = launch {
-                while (true) {
-                    val message = sendChannel.receive() as KrpcTransportMessage.StringMessage
-
-                    acting.handle(message.value.asCallMessage("1")) {
-                        fail(
-                            "Unexpected onMessageFailure call, " +
-                                    "window: ${sender.window}, collected: ${collected.size}\"",
-                            it
-                        )
-                    }.onFailure {
-                        fail(
-                            "Unexpected onFailure call, " +
-                                    "window: ${sender.window}, collected: ${collected.size}"
-                        )
-                    }.onClosed {
-                        fail(
-                            "Unexpected onClosed call, " +
-                                    "window: ${sender.window}, collected: ${collected.size}\"",
-                            it
-                        )
-                    }
-                }
-            }
-
-            val counter = Counter()
-            val printJob = launch {
-                while (true) {
-                    withContext(Dispatchers.Default) {
-                        delay(5.seconds)
-                    }
-                    println(
-                        "Collected: ${collected.size}, " +
-                                "launches: ${counter.launches.value}, " +
-                                "total: ${counter.total.value}"
-                    )
-                }
-            }
-
-            val iterations = stressIterations
-            List(iterations) {
-                launch {
-                    repeat(100) {
-                        sender.sendMessage(KrpcTransportMessage.StringMessage("Hello"))
-                        counter.total.incrementAndGet()
-                    }
-                    counter.launches.incrementAndGet()
-                }
-            }.joinAll()
-
-            while (!buffer.channel.isEmpty && sender.window != bufferSize) {
-                yield()
-            }
-
-            assertEquals(iterations * 100, collected.size)
-            actorJob.cancelAndJoin()
-            senderJob.cancelAndJoin()
-            windowJob.cancelAndJoin()
-            printJob.cancelAndJoin()
-        }
-    }
 }
 
 internal abstract class KrpcReceiveHandlerBaseTest {
